requirements.txt
alembic==0.9.9
Werkzeug==0.16.1
Flask==1.1.1
Flask-Assets==0.12
Flask-Babel==0.11.2
Flask-Bcrypt==0.7.1
Flask-Caching==1.7.1
Flask-Classy==0.6.10
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-paginate==0.5.1
Flask-Principal==0.4.0
Flask-Session==0.3.1
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
Flask-Widgets==0.4
Flask-WTF==0.14.2
gevent-websocket==0.10.1
micawber==0.5.0
MySQL-python==1.2.5
celery==4.1.0
pojson==0.7
pyOpenSSL==17.5.0
python-dateutil==2.7.2
PyYAML==3.12
speaklater==1.3
urllib3==1.22
bleach==3.1.0
ConcurrentLogHandler==0.9.1
Pillow==6.2.0
